requirements.txt
streamlit
phi
python-dotenv 
PyGithub
phidata
-e .